xlsx: fix crash when resources were missing fields
Previously the xlsx exporter blindly accessed fields that may have not been
there, due to - for instance - crash in the resource job or misformat of some
information yielded by the resource job's binary. This patch uses .get()s on
thosde dictionaries, so the worst case scenario is missing information in the
report instead of a Checkbox crash.

Fixes: LP: #1853392