fix manifest not yielding records if used on root slaves
This patch fixes a problem where slave run as root would run the
manifest job as normal user using the differential execution env.
The manifest job in the code loads all of the provider database, which
when used from snap, requires PROVIDERPATH to be set. The differential
execution environment thingy stripped that envvar. This patch add it to
the set of stored variables.